third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/events/event-listener-on-attribute-inside-shadow-dom-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: div2.createShadowRoot is not a function
This tests ensures that an event listener on an attribute node inside a shadow DOM is properly unregistered when parent element of the attribute is moved to a new document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
